rename cs2.py to cs2posts.py
